requirements.txt
langchain
langchain-community
langchain_openai
tiktoken
openai
faiss-cpu
streamlit
PyMuPDF
tqdm
python-dotenv
